[Fix] Remove models #36
