+++ paddlers/custom_models/seg/layers/__init__.py
@@ -12,4 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .farseg import FarSeg
+from .param_init import *
+from .layers_lib import *
